change to using uint32_t for channel numbers
Former-commit-id: 78c38ec0cda5e8ebaad941724b7db357a9060668

+++ ClusterFinder/fuzzyCluster_module.cc
@@ -219,7 +219,7 @@ namespace cluster{
   
     mf::LogVerbatim("Summary") << std::setfill('-') << std::setw(175) << "-" << std::setfill(' ');
     mf::LogVerbatim("Summary") << "fuzzyCluster Summary:";
-    for(unsigned int i = 0; i<ccol->size(); ++i) mf::LogVerbatim("Summary") << ccol->at(i) ;
+    for(size_t i = 0; i<ccol->size(); ++i) mf::LogVerbatim("Summary") << ccol->at(i) ;
   
     evt.put(std::move(ccol));
     evt.put(std::move(assn));
